requirements.txt
appdirs==1.4.3
attrs==18.2.0
autoflake==1.2
black==18.9b0
Click==7.0
isort==4.3.4
mypy==0.650
mypy-extensions==0.4.1
pydantic==0.18
pyflakes==2.0.0
toml==0.10.0
typed-ast==1.1.2
typing-extensions==3.7.2
ulid-py==0.0.9
